+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to MoGAAAP will be documented in this file.
 
 ### Changed
 - Update sample sheet to allow using the pipeline for already existing assemblies (#83).
+- Restructure the pipeline into three main modules: assembly, annotation and quality assessment (#91).
 
 ## [0.2.6 - 2025-02-21]