+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to MoGAAAP will be documented in this file.
 
 ### Changed
 - Update sample sheet to allow using the pipeline for already existing assemblies (#83).
+- Update Helixer to v0.3.5 which simplifies the config YAML (#92).
 
 ## [0.2.6 - 2025-02-21]