+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to MoGAAAP will be documented in this file.
 - Update SANS to version v1.0.0 (#90).
 - Restructure the pipeline into three main modules: assembly, annotation and quality assessment (#91).
 - Update Helixer to v0.3.5 which simplifies the config YAML (#92).
+- Always write scaffolded assembly as 80 character-wide fasta (#95).
 
 ## [0.2.6 - 2025-02-21]