+++ CHANGELOG.md
@@ -14,6 +14,8 @@ All notable changes to MoGAAAP will be documented in this file.
 - Restructure the pipeline into three main modules: assembly, annotation and quality assessment (#91).
 - Update Helixer to v0.3.5 which simplifies the config YAML (#92).
 - Always write scaffolded assembly as 80 character-wide fasta (#95).
+- Full restructuring of the pipeline to make modularity more logical (#89).
+- Add wrapper script to allow addition of MoGAAAP to bioconda in the future (#89).
 
 ## [0.2.6 - 2025-02-21]