+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 All notable changes to MoGAAAP will be documented in this file.
 
+## [Unreleased]
+
+### Added
+- Add wrapper script to initialise, configure and run MoGAAAP (#80).
+- Add option to use the pipeline for already existing assemblies (#83).
+- Add flye as an alternative assembler (#93).
+- Add wrapper script to download the required databases for MoGAAAP (#94).
+
+### Changed
+- Update sample sheet to allow using the pipeline for already existing assemblies (#83).
+- Update SANS to version v1.0.0 (#90).
+- Restructure the pipeline into three main modules: assembly, annotation and quality assessment (#91).
+- Update Helixer to v0.3.5 which simplifies the config YAML (#92).
+- Always write scaffolded assembly as 80 character-wide fasta (#95).
+- Full restructuring of the pipeline to make modularity more logical (#89).
+- Add wrapper script to allow addition of MoGAAAP to bioconda in the future (#89).
+
 ## [0.2.6 - 2025-02-21]
 
 ### Added